requirements.txt
PyOpenSSL==22.0.0
appdirs==1.4.4
